requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
alembic==1.12.0
email-validator==2.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
beautifulsoup4==4.12.2
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.2
coverage==7.3.1
black==23.9.1
flake8==6.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
markupsafe==2.1.3
